Clean results by fields.

Removes all records in a libgenr object that don't have specified field(s) or field(s) values.

clean_field(data, ...)

Arguments

data
a data frame object.

...
arguments specifying fields with or without values, in the format of 'field' or 'field'='value'. Separate multiple arguments with comma.

Value

cleaned data frame or FALSE if error occurs or cleaned result has 0 row.

Examples

clean_field(data=libgen_search('statistics'),'year'='2006','author')

#>     id                                    title volumeinfo
#> 1  640 Pattern Recognition and Machine Learning           
#> 9 1326        Statistics and analysis of shapes           
#>                                                           series periodical
#> 1                             Information science and statistics           
#> 9 Modeling and simulation in science, engineering and technology           
#>                      author year                          edition
#> 1     Christopher M. Bishop 2006 1st ed. 2006. Corr. 2nd printing
#> 9 Hamid Krim, Anthony Yezzi 2006                                1
#>           publisher pages language topic library issue               identifier
#> 1          Springer   749  English   183  kolxoz    17 9780387310732,0387310738
#> 9 Birkh<U+00E4>user   397  English   131  Kolxo3    21 9780817643768,0817643761
#>   issn filesize extension                              md5    crc32
#> 1       4734554       pdf 6B552B24CAE380BB656F7AAEF7F81B46 073AD785
#> 9      14052685       pdf D9738150EEE8A532D6458FE1ACA2680B F73642DE
#>                                                                                                                                                    locator
#> 1                                                     Bishop C. Pattern Recognition and Machine Learning (ISBN 0387310738)(Springer, 2006)(749s)_CsAi_.pdf
#> 9 Cs_Computer science/CsIp_Image processing/Krim H., Yezzi A. (eds.) Statistics and analysis of shapes (Birkhauser, 2006)(ISBN 0817643761)(397s)_CsIp_.pdf
#>             timeadded    timelastmodified
#> 1 2009-07-20 03:45:11 2016-03-20 07:50:50
#> 9 2009-07-20 03:45:11 2016-03-20 07:50:50
#>                                                                                                                                                                    torrent
#> 1 ZDQ6aW5mb2Q2Omxlbmd0aGk0NzM0NTU0ZTQ6bmFtZTMyOjZiNTUyYjI0Y2FlMzgwYmI2NTZmN2FhZWY3ZjgxYjQ2MTI6cGllY2UgbGVuZ3RoaTE2Nzc3MjE2ZTY6cGllY2VzMjA6KMPCrnfc2Nayr/xsFrSd+E95yRJlZQ==
#> 9 ZDQ6aW5mb2Q2Omxlbmd0aGkxNDA1MjY4NWU0Om5hbWUzMjpkOTczODE1MGVlZThhNTMyZDY0NThmZTFhY2EyNjgwYjEyOnBpZWNlIGxlbmd0aGkxNjc3NzIxNmU2OnBpZWNlczIwOixP2FqYlZYdD81SqWfwUM775pxKZWU=

clean_field(data=libgen_new_uploads(timefirst=Sys.Date()),'language'='English')

#>         id
#> 2  2448158
#> 3  2448159
#> 5  2448161
#> 6  2448162
#> 7  2448163
#> 8  2448164
#> 9  2448165
#> 10 2448166
#> 11 2448167
#> 14 2448170
#> 15 2448171
#> 17 2448173
#> 18 2448174
#> 20 2448176
#> 22 2448178
#> 23 2448179
#> 32 2448188
#> 33 2448189
#> 34 2448190
#> 35 2448191
#> 36 2448192
#> 37 2448193
#> 38 2448194
#> 39 2448195
#> 44 2448200
#> 45 2448201
#> 46 2448202
#> 48 2448204
#>                                                                                                                               title
#> 2                                                                                                           The Vegan Eggz Cookbook
#> 3                                                                                                           The Gentle Sea Cookbook
#> 5                                                    Seitan and Beyond: Gluten and Soy-Based Meat Analogues for the Ethical Gourmet
#> 6                                                                                                     Envisioning Human Geographies
#> 7                                                                             Listening to Stanley Kubrick : the music in his films
#> 8                                  The Non-Dairy Evolution Cookbook: A Modernist Culinary Approach to Plant-Based, Dairy Free Foods
#> 9                                                                NATO in the crucible : coalition warfare in Afghanistan, 2001-2014
#> 10                                                               NATO in the crucible : coalition warfare in Afghanistan, 2001-2014
#> 11                                                                         Eisenhower and the German POWs : facts against falsehood
#> 14                                                                   President Obama and Iran: Engagement, Isolation, Regime Change
#> 15                                                                                                         Sometimes a Great Notion
#> 17                                                                                          Debating Medieval Natural Law: A Survey
#> 18                                                                                                     Even a Monkey Can Draw Manga
#> 20                                                         The Autobiography of F.B.I. Special Agent Dale Cooper: My Life, My Tapes
#> 22                                                                  Evidence for Hope: Making Human Rights Work in the 21st Century
#> 23                                    Sovereignty in the south : intrusive regionalism in Africa, Latin America, and Southeast Asia
#> 32                                                      The Black Widow: My web of secrets and the truth about my murder conviction
#> 33 I Got the Show Right Here: The Amazing, True Story of How an Obscure Brooklyn Horn Player Became the Last Great Broadway Showman
#> 34                                                                                                           Stranger in the Mirror
#> 35                                                     Natasha’s Story: How a nine-year old orphan was rescued from war in Sarajevo
#> 36                                                                                                         Everything Einstein Book
#> 37                                                                              Love Gone Wrong (True Crime Box Set, Notorious USA)
#> 38                                                                                                       Atkins’ Physical Chemistry
#> 39                                                              The Gangs of Chicago: An Informal History of the Chicago Underworld
#> 44                                                        Creating Japan’s ground self-defense force, 1945-2015 : a sword well made
#> 45                                                      The Employment of English: Theory, Jobs, and the Future of Literary Studies
#> 46                                             University Reform: The Founding of the American Association of University Professors
#> 48                                                                                                 Fundamentals of Chinese Medicine
#>    volumeinfo                                             series periodical
#> 2                                                                          
#> 3                                                                          
#> 5                                                                          
#> 6                                                                          
#> 7                                                                          
#> 8                                                                          
#> 9                      Hoover Institution Press publication 703.           
#> 10                     Hoover Institution Press publication 703.           
#> 11                                                                         
#> 14                                                                         
#> 15                                                                         
#> 17                                                                         
#> 18                                                             1           
#> 20                                                                         
#> 22                      Human Rights and Crimes against Humanity           
#> 23                                                                         
#> 32                                                                         
#> 33                                                                         
#> 34                                                                         
#> 35                                                                         
#> 36                                                                         
#> 37                                                 Notorious USA           
#> 38                                                                         
#> 39                                                                         
#> 44                                                                         
#> 45                                                Cultural Front           
#> 46                                                                         
#> 48            International Standard Library of Chinese Medicine           
#>                                                                        author
#> 2                                                         Skye Michael Conroy
#> 3                                                         Skye Michael Conroy
#> 5                                                         Skye Michael Conroy
#> 6                                Paul J. Cloke; Philip Crang; Mark A. Goodwin
#> 7                                    Gengaro, Christine Lee; Kubrick, Stanley
#> 8                                                         Skye Michael Conroy
#> 9                                                            Hanagan, Deborah
#> 10                                                           Hanagan, Deborah
#> 11             Ambrose, Stephen E.; Bischof, Günter; Eisenhower, Dwight David
#> 14                                                             Raymond Tanter
#> 15                                                  Ken Kesey; Charles Bowden
#> 17                                                          Riccardo Saccenti
#> 18                                             Koji Aihara & Kentaro Takekuma
#> 20                                                                Scott Frost
#> 22                                                            Kathryn Sikkink
#> 23                                                             Coe, Brooke N.
#> 32                                                               Linda Calvey
#> 33                                                        Cy Feuer; Ken Gross
#> 34                                                      Michael Edward Little
#> 35                                                          Michael Nicholson
#> 36                                             Shana Priwer; Cynthia Phillips
#> 37                                                Caitlin Rother; Gregg Olsen
#> 38                                 Peter Atkins, Julio de Paula, James Keeler
#> 39                                                             Herbert Asbury
#> 44                                                       David Hunter-Chester
#> 45                                                             Michael Bérubé
#> 46                                           Hans-Joerg Tiede, Michael Bérubé
#> 48 Sun Guang-ren (Author), Doug Eisenstark (Author), Zhang Qing-rong (Editor)
#>    year           edition                                   publisher    pages
#> 2  2016                                                                     88
#> 3  2017                                                                    113
#> 5  2015                   CreateSpace Independent Publishing Platform      248
#> 6  2004                                              Taylor & Francis      252
#> 7  2014 1st paperback ed.                        Rowman & Littlefield      305
#> 8  2014                   CreateSpace Independent Publishing Platform      177
#> 9  2019                                      Hoover Institution Press      285
#> 10 2019                                      Hoover Institution Press      285
#> 11 1992                              Louisiana State University Press      258
#> 14 2012                                         Iran Policy Committee      141
#> 15 2006         Paperback                            Penguin Classics      715
#> 17 2016                                University of Notre Dame Press  xiv+156
#> 18 2002                                       Viz Communications Inc.      146
#> 20 1991                                                  Pocket Books      208
#> 22 2017                                    Princeton University Press viii+320
#> 23 2019                                    Cambridge University Press      228
#> 32 2019                                                  Mirror Books      336
#> 33 2010                                          Simon & Schuster Inc      304
#> 34 2014                                                 Bookwhirl.com      136
#> 35 2014                                              Apostrophe Books      200
#> 36 2003                                                    Everything      304
#> 37 2016                   Createspace Independent Publishing Platform      226
#> 38 2018              11th                     Oxford University Press      940
#> 39 2002         Paperback                       Thunder’s Mouth Press      379
#> 44 2018                                               Lexington Books      312
#> 45 1997                                                     NYU Press        0
#> 46 2015                                Johns Hopkins University Press      288
#> 48 2014                             People’s Medical Publishing House        0
#>    language topic library issue
#> 2   English                   0
#> 3   English                   0
#> 5   English                   0
#> 6   English                   0
#> 7   English                   0
#> 8   English                   0
#> 9   English                   0
#> 10  English                   0
#> 11  English    67             0
#> 14  English                   0
#> 15  English   102             0
#> 17  English   195             0
#> 18  English    59             0
#> 20  English                   0
#> 22  English   192             0
#> 23  English   193             0
#> 32  English    68             0
#> 33  English    68             0
#> 34  English    68             0
#> 35  English    68             0
#> 36  English    68             0
#> 37  English    68             0
#> 38  English   303             0
#> 39  English    68             0
#> 44  English    64             0
#> 45  English                   0
#> 46  English                   0
#> 48  English   160             0
#>                                                                    identifier
#> 2                                                                            
#> 3                                                                            
#> 5                                                                  1516860888
#> 6                                                    0340720123,9780340720127
#> 7  9780810885646,0810885646,9780810885653,0810885654,9781442244405,1442244402
#> 8                                                                  1499590423
#> 9                                                    9780817922955,0817922954
#> 10                                                   9780817922955,0817922954
#> 11                                                   9780807117583,0807117587
#> 14                                                   0979705150,9780979705151
#> 15                                                   0143039865,9780143039860
#> 17                                                   0268100403,9780268100407
#> 18                                                              1-56931-863-8
#> 20                                                                 0671744003
#> 22                                                   1400888530,9781400888535
#> 23                                                   9781108654821,1108654827
#> 32                                                   1912624591,9781912624591
#> 33                                                   0743236114,9780743236119
#> 34                                                   1618564110,9781618564115
#> 35                                                   1910167584,9781910167588
#> 36                                                   158062961X,9781580629614
#> 37                                                   1534606238,9781534606234
#> 38                                                 0198769865, 978-0198769866
#> 39                                                   1560254548,9781560254546
#> 44                                                                           
#> 45                                                 978-0814713013, 0814713017
#> 46                                                1421418266,  978-1421418261
#> 48                                                             978-7117187268
#>    issn  filesize extension                              md5    crc32
#> 2         6959457       pdf dff2ecd9e49a48f79b23007dc357e663 BA9D6248
#> 3         4850202       pdf 69130c545617145301fa6b485a02b70b C65F0F58
#> 5        14911134       pdf 6663a885f7529364ad907832454caf64 6C5346F8
#> 6         1427684       pdf 933538132e60692b2a1111f63eeaeec3 2D96A9B4
#> 7         1222796      epub 4db81b73c5f7ae83b7444f7933eab470 A34C849D
#> 8         7577127       pdf d46e3df26751b9f8df7660708010455b F896EC67
#> 9         5744927      epub 46563ee3bb38a64845ac013aa193cc05 DAC6E9C4
#> 10        3821861       pdf 0548a97225e6721480155064cf74f6f3 9ACAE391
#> 11       24690770       pdf 5af750d25945b28ba5c0c5227cc059b1 20D0E00A
#> 14        7318117       pdf db24631a920162b9c2a1a0479ce6f148 C5248B9D
#> 15         800718      epub 547de1a3b5d0793c5b345cb7ad90b3cb 20BA1ACB
#> 17        1157162       pdf 32c98b79253288790f0b891b3418dd1b 2A6E37B2
#> 18       27822920       pdf 9768123f1c149a7332cbe7e09d3e8df0 D9001355
#> 20       20913205       pdf 0f89ec38bcfcf68913d5a2e9afcbb5da 413240FA
#> 22        6724684       pdf 212a0e2e8694375a8f611e23a0bf73a4 DD943E7A
#> 23        1159679       pdf 478ad4cba118cbb5bcf94425d42297b1 85A92A29
#> 32         310474      epub c4a420aecc79ba93456f926af0941ea1 F2A78558
#> 33        1424370      epub ea8e7c1c471ca287522c5e33ad66f7f5 00E60C36
#> 34         417715      epub 2d3281c720e94d248375681e977a8232 D34A55D9
#> 35        2321787      epub e1c04ee01c96e1ef7050ea903e0a7a58 46A8DE24
#> 36         551024      epub 897f9dcdcc61172f086be620ae70d55a FE80AA11
#> 37         497510      epub a6296df3ab9d362d71c699e4846812bb 08E0F3C1
#> 38      100914973       pdf 863e2263e0201f6bacd22cf37d7d0db3 27D34E6E
#> 39       10672161      epub 9f74820099640ece2b8b92df779398f9 7B43B6C7
#> 44       10662979       pdf 4f81b9a56d8464da4be08abb3262ef7c 38DE0CFB
#> 45        2046043      epub b94cbaf750987b0bc44fabdef65a5c15 0ACCC694
#> 46       11747417       pdf 79d1bbacfc9f65b3372552e2e052dc20 B5727616
#> 48       16815698      epub d301b21756e8126d6d736f69503f8191 A8D6273D
#>                                                                                                          locator
#> 2                                                                                               eggzcookbook.pdf
#> 3                                                                                    the_gentle_sea_cookbook.pdf
#> 5                                                           Seitan and Beyond Cookbook - Skye Michael Conroy.pdf
#> 6                          Envisioning Human Geographies - Cloke, Paul; Crang, Philip; Goo 978-0-340-72012-7.pdf
#> 7                                                            Listening to Stanley Kubrick 978-0-8108-8564-6.epub
#> 8                                                                            thenondairyevolutioncookbookpdf.pdf
#> 9                                                                    NATO in the Crucible 978-0-8179-2295-5.epub
#> 10                                                                    NATO in the Crucible 978-0-8179-2295-5.pdf
#> 11                                                                                               EPOW_nometa.pdf
#> 14                                                                President Obama and Iran 978-0-9797051-5-1.pdf
#> 15                                                                     Sometimes a Great Notion - Ken Kesey.epub
#> 17                                                          Saccenti_Debating Medieval Natural Law. A Survey.pdf
#> 18                                                                              Even A Monkey Can Draw Manga.pdf
#> 20                                          The autobiography of F.B.I. special agent Dale Cooper  my life, .pdf
#> 22                                   Sikkink_Evidence for Hope. Making Human Rights Work in the 21st Century.pdf
#> 23 Coe (2019) - Sovereignty in the South. Intrusive Regionalism in Africa, Latin America, and Southeast Asia.pdf
#> 32                                                           The Black Widow_ The true crime - Linda Calvey.epub
#> 33                                                         I Got The Show Right Here by Cy Feuer, Ken Gross.epub
#> 34                                                   Stranger in the Mirror<U+00BB>by Michael Edward Little.epub
#> 35                                                                      Natashas Story by Michael Nicholson.epub
#> 36                                           The Everything Einstein Book by Shana Priwer, Cynthia Phillips.epub
#> 37                                                                                               CaitRotLGW.epub
#> 38                                                                           Atkins__039; Physical Chemistry.pdf
#> 39                                                                    The Gangs of Chicago - Herbert Asbury.epub
#> 44                                                                                             Untitled.FR12.pdf
#> 45                                                         The Employment of English_ Theo - Michael Berube.epub
#> 46                                                        University Reform_ The Founding - Hans-Joerg Tiede.pdf
#> 48                                                          Fundamentals of Chinese Medicine - Guangren Sun.epub
#>              timeadded    timelastmodified torrent
#> 2  2019-12-11 00:03:08 2019-12-11 00:03:09      NA
#> 3  2019-12-11 00:05:27 2019-12-11 00:05:28      NA
#> 5  2019-12-11 00:11:26 2019-12-11 00:11:27      NA
#> 6  2019-12-11 00:12:09 2019-12-11 00:12:11      NA
#> 7  2019-12-11 00:14:50 2019-12-11 00:14:51      NA
#> 8  2019-12-11 00:15:46 2019-12-11 00:15:47      NA
#> 9  2019-12-11 00:17:08 2019-12-11 00:17:09      NA
#> 10 2019-12-11 00:19:06 2019-12-11 00:19:06      NA
#> 11 2019-12-11 00:19:23 2019-12-11 00:29:02      NA
#> 14 2019-12-11 00:25:10 2019-12-11 00:25:11      NA
#> 15 2019-12-11 00:28:22 2019-12-11 00:28:24      NA
#> 17 2019-12-11 00:49:54 2019-12-11 00:49:55      NA
#> 18 2019-12-11 00:57:37 2019-12-11 00:57:38      NA
#> 20 2019-12-11 01:03:12 2019-12-11 01:03:14      NA
#> 22 2019-12-11 01:16:32 2019-12-11 01:16:33      NA
#> 23 2019-12-11 01:49:28 2019-12-11 01:49:29      NA
#> 32 2019-12-11 03:09:42 2019-12-11 03:09:43      NA
#> 33 2019-12-11 03:12:59 2019-12-11 03:13:00      NA
#> 34 2019-12-11 03:15:18 2019-12-11 03:15:19      NA
#> 35 2019-12-11 03:17:37 2019-12-11 03:18:29      NA
#> 36 2019-12-11 03:24:00 2019-12-11 03:24:01      NA
#> 37 2019-12-11 03:31:06 2019-12-11 03:31:08      NA
#> 38 2019-12-11 03:34:40 2019-12-11 03:34:44      NA
#> 39 2019-12-11 03:37:14 2019-12-11 03:37:15      NA
#> 44 2019-12-11 04:18:42 2019-12-11 04:18:43      NA
#> 45 2019-12-11 05:08:58 2019-12-11 05:08:59      NA
#> 46 2019-12-11 05:17:16 2019-12-11 05:17:18      NA
#> 48 2019-12-11 06:00:24 2019-12-11 06:00:25      NA